requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
jupyter=1.0.0=py37_7
jupyter_client=5.3.4=py37_0
jupyter_console=6.1.0=py_0
jupyter_core=4.6.1=py37_0
matplotlib=3.1.0=py37h5429711_0
notebook=6.0.3=py37_0
numpy=1.18.1=py37h95a1406_0
numpy-base=1.18.1=py37h2f8d375_1
python=3.7.6=h0371630_2
python-dateutil=2.8.1=py_0
pytorch=1.2.0=py3.7_cuda9.2.148_cudnn7.6.2_0
scikit-learn=0.22.1=py37hcdab131_1
scipy=1.4.1=py37h921218d_0
torchvision=0.4.0=py37_cu92